Use poetry to manage dependencies and packaging
Poetry is a tool to manage depencies in Python. It also
allows to create wheel and source packages.

When poetry is in use, it creates a 'poetry.lock' file
which stores all the depencencies which their specific
versions needed to use the package. Therefore, it allows
to have deterministic deployments.

After adding poetry, setup files ('setup.py', 'requirements.txt'
and so on) are not needed anymore.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>