[changelog] Replace pull request field for issue in changelog entries
The initial idea of the pull request field in a changelog
entry was to know which PR introduced the change. The problem
is the reference is unknown until the PR is generated, so
developers will need to modify the entry and the PR later
to include the reference. Therefore, we considered it is
better to include a reference to an issue which is more
generic and can be known in advance.

Take into account you will need to update your unreleased
changelog entries if you want to use this new version of
the tools.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>