[changelog] Validate the format of the yml file
Fixes #17.

This code validates the format of the yml file and gives info
about the errors, if any. It gives an option to edit the file
incase it if has any mistakes.

This feature is necessary to avoid any mistakes by the user
while editing the yml file using the changelog tool.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>