[semverup] Option to create release candidates
This commit includes a new argument to semverup to create
prereleases versions.

If the option `--prerelease` is included and the current version
is not a release candidate, it will create a new rc version based
on files or `--bump-version` argument. If the current version
is a release candidate, it will increase its prerelease part.

When the version is a prerelease, and `--prerelease` is not specified,
it will remove the prerelease part and generate the final version.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>