[notes] Changelogs location updated between versions
'notes' move the changelog entries to a new directory when they
are used for creating a release notes file. Notes has a now a
new argument '--pre-release', when it is defined, it ignores
'unreleased/processed' notes for creating new release notes,
when it is not passed, it includes 'unreleased/processed' in
release notes.

'semverup' now updates the version depending on the notes that
are in 'unreleased' directory and ignores 'unreleased/processed'.

'publish' removes the notes at unreleased/processed because are
the notes that are already processed and ignores the yml files
at unreleased/*.yml

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>