Remove mix_stderr parameter from CliRunner instances
This commit removes the mix_stderr=False parameter from all
instances of CliRunner in the test suite.

The mix_stderr was required to separate the standard error
from the standard output in the test results. However, in
the latest version of the package this argument was removed
and now always collect stderr output and never raise an
exception

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>